requirements.txt
arabic-reshaper==2.1.4
asgiref==3.5.2
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
cssselect2==0.7.0
Django==4.1.2
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-filter==22.1
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-weasyprint==2.2.0
django-widget-tweaks==1.4.12
django-xhtml2pdf==0.0.4
fonttools==4.38.0
future==0.18.2
html5lib==1.1
idna==3.4
jdatetime==4.1.0
lxml==4.9.1
oscrypto==1.3.0
Pillow==9.2.0
pycparser==2.21
pydyf==0.5.0
pyHanko==0.15.0
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
pyphen==0.13.0
python-bidi==0.4.2
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
tzdata==2022.5
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
weasyprint==57.0
webencodings==0.5.1
xhtml2pdf==0.2.8
zopfli==0.2.1
